'backouted' a few changes, made AStar3D::Container::update behave correctly
Backouted some unneccesary visibility changes and made the hashmap in
AStar3D::Container store Nodes instead of Node*s. That may increase memory
usage, but gets rid of some valgrind errors.

+++ source/actions/changeview.h
@@ -29,7 +29,7 @@
 
 class ChangeView : public GameAction {
    public:
-      typedef map<MapCoordinate,Uint64> ViewState;
+      typedef map<MapCoordinate,int> ViewState;
       ChangeView( GameMap* gamemap, const ViewState& state );
